refactor(oga): extract storage domain into dedicated internal/storage package
