+++ oga/internal/handler.go
@@ -221,51 +221,3 @@ func (h *OGAHandler) HandleReviewApplication(w http.ResponseWriter, r *http.Requ
 		"message": "Application reviewed successfully",
 	})
 }
-
-// HandleGetUploadURL returns a download URL for a file stored in the main
-// backend's upload service.  The OGA frontend calls this endpoint and expects
-// a JSON response with {download_url, expires_at} so that the FileControl
-// component can set the <a href> to the resolved URL.
-func (h *OGAHandler) HandleGetUploadURL(w http.ResponseWriter, r *http.Request) {
-	key := r.PathValue("key")
-	if key == "" {
-		WriteJSONError(w, http.StatusBadRequest, "key is required")
-		return
-	}
-	metadata, err := h.service.GetDownloadURL(r.Context(), key)
-	if err != nil {
-		slog.ErrorContext(r.Context(), "failed to get download URL from backend", "key", key, "error", err)
-		WriteJSONError(w, http.StatusInternalServerError, "failed to get download URL")
-		return
-	}
-
-	WriteJSONResponse(w, http.StatusOK, metadata)
-}
-
-// HandleCreateUpload prepares an upload by requesting an upload URL from the main backend
-func (h *OGAHandler) HandleCreateUpload(w http.ResponseWriter, r *http.Request) {
-	// TODO: Add Authentication & Authorization middleware
-	// Access must be restricted to authorized OGA officers to prevent unauthorized users
-	// from generating proxy pre-signed upload URLs. Introduce a configuration flag (e.g. OGA_DISABLE_AUTH)
-	// to make bypassing explicit for specific environments
-	if r.Method != http.MethodPost {
-		WriteJSONError(w, http.StatusMethodNotAllowed, "Method not allowed")
-		return
-	}
-
-	r.Body = http.MaxBytesReader(w, r.Body, h.MaxRequestBytes)
-	var body json.RawMessage
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		WriteJSONError(w, http.StatusBadRequest, "Invalid request body")
-		return
-	}
-
-	result, err := h.service.CreateUploadURL(r.Context(), body)
-	if err != nil {
-		slog.ErrorContext(r.Context(), "failed to create upload URL", "error", err)
-		WriteJSONError(w, http.StatusInternalServerError, "Failed to create upload URL")
-		return
-	}
-
-	WriteJSONResponse(w, http.StatusOK, result)
-}
